@@ -1,6 +1,6 @@
 module github.com/pruh/api
 
-go 1.18
+go 1.19
 
 require (
 	github.com/golang/glog v1.0.0
