@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -127,10 +127,10 @@ func TestTelegramControllerSendMessage(t *testing.T) {
 	assert := assert.New(t)
 
 	for _, testData := range testsData {
-		t.Logf("tesing %+v", testData)
+		t.Logf("tesing %s", testData.description)
 
 		controller := Controller{
-			Config: NewConfigSafe(strPtr("8080"), strPtr("1"), testData.defaultChatID, nil, nil, nil),
+			Config: NewConfigSafe(strPtr("8080"), strPtr("1"), testData.defaultChatID, nil, nil, nil, nil, nil, nil),
 			HTTPClient: &MockHTTPClient{
 				do: func(req *http.Request) (*http.Response, error) {
 					if !testData.telegramShouldBeCalled {
@@ -167,7 +167,7 @@ func TestTelegramControllerSendMessage(t *testing.T) {
 }
 
 func formatBody(w *httptest.ResponseRecorder) string {
-	bodyBytes, err := ioutil.ReadAll(w.Body)
+	bodyBytes, err := io.ReadAll(w.Body)
 	if err != nil {
 		panic(fmt.Sprintf("Error while reading body: %s", err))
 	}
