@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"strings"
 	"time"
@@ -106,7 +105,7 @@ func (c *Controller) Get(w http.ResponseWriter, r *http.Request) {
 // Create creates a new notification.
 func (c *Controller) Create(w http.ResponseWriter, r *http.Request) {
 	var notification Notification
-	body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1024*1024))
+	body, err := io.ReadAll(io.LimitReader(r.Body, 1024*1024))
 	if err != nil {
 		glog.Errorf("Error reading request. %s", err)
 		http.Error(w, "Error reading request.", http.StatusInternalServerError)
