@@ -3,10 +3,11 @@ package config
 import (
 	"encoding/json"
 	"errors"
-	"github.com/golang/glog"
 	"net"
 	"os"
 	"strconv"
+
+	"github.com/golang/glog"
 )
 
 // Configuration contrains configuration parameters.
@@ -19,6 +20,10 @@ type Configuration struct {
 
 	MongoUsername *string
 	MongoPassword *string
+
+	OmadaUrl      *string
+	OmadaUsername *string
+	OmadaPassword *string
 }
 
 // NewFromEnv creates new configuration from environment variables.
@@ -31,12 +36,18 @@ func NewFromEnv() (*Configuration, error) {
 	mongoUsername := ptrOrNil(os.LookupEnv("MONGO_INITDB_ROOT_USERNAME"))
 	mongoPassword := ptrOrNil(os.LookupEnv("MONGO_INITDB_ROOT_PASSWORD"))
 
-	return NewFromParams(port, botToken, chatID, apiCreds, mongoUsername, mongoPassword)
+	omadaUrl := ptrOrNil(os.LookupEnv("OMADA_URL"))
+	omadaUsername := ptrOrNil(os.LookupEnv("OMADA_USERNAME"))
+	omadaPassword := ptrOrNil(os.LookupEnv("OMADA_PASSWORD"))
+
+	return NewFromParams(port, botToken, chatID, apiCreds, mongoUsername, mongoPassword,
+		omadaUrl, omadaUsername, omadaPassword)
 }
 
 // NewFromParams creates new configuration from arguments.
 func NewFromParams(port *string, boToken *string, defaultChatID *string,
-	apiV1Credentials *string, mongoUsername *string, mongoPassword *string) (*Configuration, error) {
+	apiV1Credentials *string, mongoUsername *string, mongoPassword *string,
+	omadaUrl *string, omadaUsername *string, omadaPassword *string) (*Configuration, error) {
 	var conf Configuration
 	if port == nil || *port == "" {
 		return nil, errors.New("port should not be empty")
@@ -68,6 +79,12 @@ func NewFromParams(port *string, boToken *string, defaultChatID *string,
 	if mongoUsername != nil && mongoPassword != nil {
 		conf.MongoUsername = mongoUsername
 		conf.MongoPassword = mongoPassword
+	}
+
+	if omadaUrl != nil {
+		conf.OmadaUrl = omadaUrl
+		conf.OmadaUsername = omadaUsername
+		conf.OmadaPassword = omadaPassword
 	}
 
 	return &conf, nil
