@@ -14,6 +14,11 @@ func TestConfiguraionLoading(t *testing.T) {
 		botToken      *string
 		defaultChatID *string
 		credsMap      *map[string]string
+		mongoUsername *string
+		mongoPassword *string
+		omadaUrl      *string
+		omadaUsername *string
+		omadaPassword *string
 		expectError   bool
 	}{
 		{
@@ -87,6 +92,30 @@ func TestConfiguraionLoading(t *testing.T) {
 			credsMap:      &map[string]string{},
 			expectError:   false,
 		},
+		{
+			description:   "mongo credentials parsing",
+			port:          ptr("1234"),
+			botToken:      ptr("botToken"),
+			mongoUsername: ptr("mongoUsername"),
+			mongoPassword: ptr("mongoPassword"),
+			expectError:   false,
+		},
+		{
+			description: "omada url",
+			port:        ptr("1234"),
+			botToken:    ptr("botToken"),
+			omadaUrl:    ptr("abc"),
+			expectError: false,
+		},
+		{
+			description:   "omada url with credentials",
+			port:          ptr("1234"),
+			botToken:      ptr("botToken"),
+			omadaUrl:      ptr("abc"),
+			omadaUsername: ptr("omada username"),
+			omadaPassword: ptr("omada password"),
+			expectError:   false,
+		},
 	}
 
 	assert := assert.New(t)
@@ -94,10 +123,12 @@ func TestConfiguraionLoading(t *testing.T) {
 	for _, testData := range testsData {
 		t.Logf("testing %+v", testData.description)
 
-		conf, err := NewConfig(testData.port, testData.botToken, testData.defaultChatID, testData.credsMap, nil, nil)
+		conf, err := NewConfig(testData.port, testData.botToken, testData.defaultChatID,
+			testData.credsMap, testData.mongoUsername,
+			testData.mongoPassword, testData.omadaUrl, testData.omadaUsername, testData.omadaPassword)
 
 		if !testData.expectError && err != nil {
-			assert.Fail("Config load should not return error: %s", err.Error)
+			assert.Fail("Config load should not return error: " + err.Error())
 			continue
 		}
 
@@ -113,6 +144,9 @@ func TestConfiguraionLoading(t *testing.T) {
 		assert.Equal(testData.port, conf.Port, "Port is not correct")
 		assert.Equal(testData.botToken, conf.TelegramBoToken, "Bot token is not correct")
 		assert.Equal(testData.credsMap, conf.APIV1Credentials, "Credentials is not correct")
+		assert.Equal(testData.mongoUsername, conf.MongoUsername, "Mongo username is not correct")
+		assert.Equal(testData.mongoPassword, conf.MongoPassword, "Mongo password ID is not correct")
+		assert.Equal(testData.omadaUrl, conf.OmadaUrl, "Omada Controller ID is not correct")
 	}
 }
 
